[Tests] Always run test.manager CLI via wrapper
One test was using its own mechanism to launch the CLI under the guise
of its use case being "different enough". This increases maintenance
overhead and the risk of a divergent approach.

This commit updates execute_cli to allow None for the fixture file path,
which causes the argument to be omit.

Signed-off-by: Tom Cowland <tom@foundry.com>