[Tests] Fix python used in test.manager subprocesses
Using "python" results in whatever python is on $PATH being used.
In the ctest invocation (which uses a venv python, without activating
the venv), this is different to the executable running the tests.

This results in errors such as:

   Error while finding module specification for 'openassetio.test.manager'
   (ModuleNotFoundError: No module named 'openassetio')

As the system python paths are not configured to include the built
modules et. al.

Signed-off-by: Tom Cowland <tom@foundry.com>