[Build] Ensure ctest runs pytest from project root
Previously, we ran from the tests directory. As this contains a module
structure that exactly mirrors the main openassetio package, any imports
end up targeting the tests modules, not the openassetio ones, due to
'.' being ahead of everything else on sys.path.

This results in (confusing) errors such as:

  No module named openassetio.test.manager.__main__;
  'openassetio.test.manager' is a package and cannot be directly executed.

Running from the root also allows us to include the resources directory
as well should we wish.

Signed-off-by: Tom Cowland <tom@foundry.com>