[Core] Migrate `defaultEntityReference` to use trait sets
Updates the signature for `defaultEntityReference` to take a set of trait
IDs rather than a `Specification`. This removes the ambiguity around
whether or not the method should consider the properties of the traits
vs their identifiers.

This method is meant as a high-level starting point for some "kind" of
entity, before they exist, so it should not be affected by the data of
any particular entity itself.

Closes #281.

Signed-off-by: Tom Cowland <tom@foundry.com>