[Docs] Add Doxygen alias for C++ namespace prefix
For #292 we migrated some `ManagerInterface` methods to C++, including
their docstrings. Doxygen then renders the docstrings under the
"invisible" `inline namespace OPENASSETIO_VERSION`, for better or for
worse.

Rather than reproduce the version tag verbatim for every reference,
causing a maintenance burden, add a Doxygen `ALIAS` command of `fqref`
for "fully-qualified @ref", which will prefix a given argument with
the top-level namespaces.

Also use an environment variable to supply the `OPENASSETIO_VERSION`
substitution in Doxygen, and set this in the `doc/Makefile`.

Signed-off-by: David Feltell <david.feltell@foundry.com>