[Core] Migrate `getRelatedReferences` to use trait sets
Updates the signature for `getRelatedReferences` to take a set of trait
IDs rather than a `Specification`. This removes the ambiguity around
whether or not the method should consider the properties of the traits
vs their identifiers.

This argument is meant to indicate the "type" of entities returned,
trait properties form the data of a specific entity.

Closes #282.

Signed-off-by: Tom Cowland <tom@foundry.com>