[Core] Remove attributes API
As part of #286, the attributes mechanism goes away in favour of
registering/resolving traits and their properties. This change is
discussed in more detail in DR008.

This interim commit removes the attributes API, and updates the
surrounding documentation.

The only documentation left un-altered is the "Entities, Specifications
and Attributes" page, which is tackled independently under #249.

Closes #307

Signed-off-by: Tom Cowland <tom@foundry.com>