[Lint] Clang-Tidy rules for C/C++ tests
When writing C API tests for #294 we have opted to use Catch2 and the
Trompeloeil mocking framework.

Catch2 triggers `readability-function-cognitive-complexity` due to the
size of its macros.

Trompeloeil triggers `readability-identifier-naming` due to the use of
placeholder variables `_1`, `_2`, etc.

So add a `.clang-tidy` config override in the `tests` directory. This
works to disable `readability-function-cognitive-complexity`.

However, the `tests`-specific `.clang-tidy` doesn't work for configuring
a `VariableIgnoredRegexp` for `readability-identifier-naming`. This has
been reported upstream. In the meantime, just add
`VariableIgnoredRegexp` to the global config, where it works.

Signed-off-by: David Feltell <david.feltell@foundry.com>