[Build] Add `clang-format` check to build
For #251, add an optional `clang-format` check as part of the CMake
build, implemented in much the same way as the `cpplint` check. The main
difference is that `clang-format` has no built-in support for
recursively checking all files in a directory tree, so we must build
the list of files ourselves.

Signed-off-by: David Feltell <david.feltell@foundry.com>