[Core] Migrate `preflight` to use trait sets
Updates the signature for `preflight` to take a set of trait
IDs rather than a `Specification`. This removes the ambiguity around
whether or not the method should consider the properties of the traits
vs their identifiers.

This argument is meant to indicate the "type" of entities that are about
to have their data generated. Supplying the properties for a trait here
makes no sense, as that data _is_ the data of the entity, and so the
host can't provide it yet, as its not been made!

Closes #283.

Signed-off-by: Tom Cowland <tom@foundry.com>