[Docs] Fix README ASWF Docker build instructions
The addition of C++ tests for #294 means that we now have C++ library
dependencies that are not present in the ASWF Docker image, specifically
the Catch2 and Trompeloeil testing libraries. This means we can no
longer build and run all `ctest` targets from a container.

Upcoming work will add some build tooling for those who wish to run
tests via the Docker image.

In the meantime, reduce the scope of the instructions in the README to
building and installing the compiled artifacts only. I.e. we lose the
creation of a Python environment and the installation and execution of
Python and C/C++ tests.

Signed-off-by: David Feltell <david.feltell@foundry.com>