[Examples] Fix asserts in BAL tests
The infinite confusion that is using both `unittest` and `pytest`, and
then calling `unittest` tests from `pytest` tests, leads to programmers
like me, make mistakes like this. 🙃

There is a question as to whether we should do something about this,
but that is another story for when we have more time.

Signed-off-by: Tom Cowland <tom@foundry.com>