[Lint] Check C/C++ tests for linter issues
Whilst working on #290 it became apparent that we had not enabled C/C++
linter checks in the `tests` directory (we've needed them since #312).
We also forgot to add `clang-format` to the CMake status report in #313.

So add `tests` directory to CppLint and ClangFormat checks. Hence fix
CppLint issues with `#include` order in the tests sources.

Move `CPPLINT.cfg` to top-level so it also covers the `tests` directory.

Add some suppressions to `CPPLINT.cfg` (see comments).

Signed-off-by: David Feltell <david.feltell@foundry.com>