[Core] Validate specification trait sets passed to `register`
Ensures that the caller of `Manager.register` is adhering to the API
contract that all supplied specifications should share a trait set.

There is potentially a cost to this validation, but we consider
`register` to be a lower volume call than say, `resolve`, and so this
overhead is negligible, and the value in ensuring coherence for the
manager implementation is worth the cost.

Closes #328.

Signed-off-by: Tom Cowland <tom@foundry.com>