[Core] C API InfoDictionary `typeOf` function
Part of #290. Add an introspection function allowing the type of an
existing entry to be queried, so that the correct accessor function
can be chosen.

A C enum mirrors the types available in the underlying C++
`std::variant`. We could have used string literals or integral constants
instead of an enum. However, the available types cannot be extended, and
we expect the list of types to be parsed by language binding
generators/FFIs. So an enum was chosen to provide a more concrete and
exhaustive list.

Signed-off-by: David Feltell <david.feltell@foundry.com>