[Core] C API InfoDictionary remove suite
Part of #290.

Previously, symmetry was assumed between the mechanism for plugins
providing C API functions to OpenAssetIO, and OpenAssetIO provided C API
wrappers.  Specifically, this meant the C API wrappers were implemented
as a suite of function pointers, in a similar way to plugins.

However, such function pointer suites are not idiomatic C style, and are
increase boilerplate when developing plugins and bindings to other
languages that interface with the C API.

So redesign the `InfoDictionary` C API to use a series of standalone
functions, allowing for simple linkage by users.

Signed-off-by: David Feltell <david.feltell@foundry.com>