[Core] Add CManagerInterface::info
Part of #290. Add the `ManagerInterface` C API `info` suite function and
its C++ `CManagerInterface` wrapper.

The `info` suite function signature expects to be provided a C handle to
a pre-existing `InfoDictionary` as it's output parameter. The plugin is
expected to populate the `InfoDictionary` with the appropriate entries,
using its C API.

Signed-off-by: David Feltell <david.feltell@foundry.com>