[Core] C API use enum for error codes
Part of #290. For the C API we must return error codes rather than
throw exceptions. These error codes were previously designed as `int`
values, with some well-defined constants, but otherwise allowing
flexibility of the implementation to respond with any code.

However, this has the disadvantage that callers cannot be certain in
advance of the possible classes of error that can happen. Using an enum
of well-defined common error codes means cases can be guarded with
appropriate logic, with a reasonable expectation that no value outside
the enum range will be returned.

Enums also have advantages in terms of static analysis and codegen,
aiding devs and, potentially, FFIs.

Signed-off-by: David Feltell <david.feltell@foundry.com>