[Core] Add C++ Manager class
For #345 we need a C++ implementation of the most basic `Manager`
functionality, specifically `identifier`, `displayName` and `info`, to
match the currently implemented methods of `ManagerInterface`.

So add a C++ `Manager` class with implementation of the three member
functions to forward to the held `ManagerInterface`.

Skip the implementation of auditing (the main reason for the `Manager`
wrapper to exist) for now. Also don't hold a `HostSession` yet, since it
has not yet been migrated to C++, and is not required for the three
implemented methods.

Assume the `ManagerInterface` instance injected into the `Manager` is
also held elsewhere (e.g. by the plugin), hence use a `shared_ptr`.
Typedef `shared_ptr` as `openassetio::SharedPtr` simply as a good
practice.

Signed-off-by: David Feltell <david.feltell@foundry.com>