[Core] Add Python bindings for C++ Manager
For #345 the Python `Manager` must inherit from the C++ `Manager` for
now, until we complete the transition to a fully C++ (but Python-bound)
`Manager`.

So bind the C++ `Manager` to Python and inherit from it, relying on the
bound C++ implementation of `identifier`, `displayName` and `info`
methods.

Update tests to work, now that a "real" `ManagerInterface` instance is
required for the C++ `Manager` constructor:
* Many tests failed where a `Manager` must be constructed because of the
  stricter type checking that pybind11 imposes on the `ManagerInterface`
  provided to the `Manager` constructor. So make the
  `ValidatingMockManagerInterface` a common fixture, and use wherever
  tests were failing.
* Don't hoist all of `_openassetio` into the top-level, since its
  submodules conflict with packages, causing import errors.

Signed-off-by: David Feltell <david.feltell@foundry.com>