[Build] Allow `openassetio-python-venv` to build when tests are disabled
As tests require extra build-time dependencies (Catch2, Trompeloeil),
it limits the use of some off-the-shelf build environments such as the
ASWF docker containers. Moving the venv build outside of the test-only
target set allows these environments to be used to produce a working
deployment without the test dependencies.

For consumers of OpenAssetIO, this is helpful, as they are most likely
using a sanctioned tag/release and so don't need to test the API
itself.

Signed-off-by: Tom Cowland <tom@foundry.com>