[Core] Add common HandleConverter specialisations
To avoid redeclaring `HandleConverter` specialisations in code and
tests, add specialisations under the `handles` namespace, with a header
file structure mirroring the namespace structure of the associated type.

This also has the advantage of documenting whether a given handle is
meant to contain a raw pointer or a `SharedPtr` (as is the case for
`managerAPI_ManagerInterface_h`).

Also do some misc cleaning up of namespaces and typedefs.

Signed-off-by: David Feltell <david.feltell@foundry.com>