[Core] Disambiguate ManagerInterface C plugin API
In the C API a `ManagerInterface` can mean two things
* The implementation of the functionality of a `ManagerInterface`
  by a C manager plugin, consisting of a struct of function
  pointers and an opaque handle (set to whatever value the plugin
  desires).
* The C representation (handle) of a C++ `ManagerInterface` instance
  within OpenAssetIO.

So far we have used the same handle type for both cases, when they are
distinct use cases that each require their own handle.

So split the `ManagerInterface` opaque handle into two.

Rename one to call out that it is part of the C plugin system, i.e.
`managerAPI_CManagerInterface`. Also make it clear that the class that
wraps this handle and its associated function pointer suite is an
adapter class delegating to the C API, i.e. `CManagerInterfaceAdapter`.

Leave the name of the other handle as it was, i.e.
`managerAPI_ManagerInterface`, which is the handle to the C++ class
used in the construction of a `Manager`.

Signed-off-by: David Feltell <david.feltell@foundry.com>