[Build] Fix python site-packages path computation
We had been using a clever trick to work out the relative path between
Python_EXECUTABLE and Python_SITEARCH. This had issues in certain
situations, where they may not share a common Python installation root.

This may be due to symlinks making paths _look_ different, or just be
a by product of the specific deployment.

This is really just trying to work out the right platform-specific
sub-directory that is always under --install-dir to put the package in.
So, for now, we revert to a simpler approach that makes a high-level
switch between the well-known defaults for our specific target platforms.

We may need to revisit this again, but it seems more robust in the sort
term.

Signed-off-by: Tom Cowland <tom@foundry.com>