[Build] Unset .so version for Python extension
CMake's `SOVERSION` corresponds to the "compatibility version" for
dynamic libraries on macOS, and `VERSION` corresponds to "current
version". This appears to be incompatible with `.so` libs (required for
Python), as opposed to `.dylib`s.

We already override the `_openassetio.so` library name to not have the
version suffix anyway, so no harm in just removing them from the
target properties to prevent CMake adding the incompatible linker flags.

Signed-off-by: David Feltell <david.feltell@foundry.com>