[Lint] Default to trim whitespace in .editorconfig
The reason for not doing this before is that some languages may have
significant trailing whitespace. In particular, a double trailing
whitespace in markdown forces a line break. However, there are better
alternatives to achieving the same effect. The amount of trailing
whitespace that slips through is far more of a problem.

Signed-off-by: David Feltell <david.feltell@foundry.com>