[Core] Decouple symbol namespace version from release version
We previously used the major and minor release version components to
define the C/C++ symbol namespace. This had the downside that it would
change with every minor version, forcing a re-compile even if the ABI
had not altered.

There are many cases in which either of these two version components may
be bumped without the ABI being affected (python only changes, feature
additions, etc.).

The new approach freezes the namespace version to the last major version
where the ABI changed (as an ABI break would always require a major
release). Hopefully this will avoid the need for recompilation in many
situations and reduce the maintenance overhead.

Closes #377.

Signed-off-by: Tom Cowland <tom@foundry.com>