[Resources] Add `openassetio-codegen` tool
Adds a tool to automatically generate Trait and Specification view
classes from a simple YAML description format.

This avoids laborious and error-prone manual authoring for each
supported language.

The tool can be used on the command line, or as a python module for
programatic/custom code generation.

This initial version only supports Python 3.9+ generation, other
languages will follow in due course.

The tool itself has no runtime dependencies on OpenAssetIO, so it can
be used to autogenerate core API classes themselves.

We chose to use Jinja for generation, however this can be changed
per-language if required.

The high-level code generation process is broken down to the following
steps:

 - Load description from YAML.
 - Validate the resulting data structure using JSON Schema.
 - Build an intermediate code generation declaration model, decoupled
   from the source description.
 - Pass this declaration to one or more generators.

The declaration is centered around the idea of a package of traits
and specifications. Each package:

 - Has a universally unique identifier specified by the author.
 - Can contain any number of Traits and/or Specifications.
 - Conceptually similar Traits/Specifications are grouped under a
   simple single-level namespace structure.
 - The resulting Trait IDs follow the form: <package>:<namespace>.<name>
   to ensure they are universally unique without constraining their
   individual names.
 - Specification TraitSets can reference traits from their own package,
   or another by its ID.

Each of the generation steps are exposed via public APIs in the codegen
package.

Closes #415

Signed-off-by: Tom Cowland <tom@foundry.com>