[Core] Migrate Host/HostInterface to C++
Part of #331. Migrate `Host` and `HostInterface` classes to C++ with
Python bindings.

Remove `HostInterface.py`, since it offers nothing beyond the C++
implementation. However, retain `Host` since it has some methods that
are yet to be migrated.

Add a `MockHostInterface` class, following a similar design to the
`ValidatingMockManagerInterface`, so that pybind type assertions work.
Hence fix existing tests to use this.

Signed-off-by: David Feltell <david.feltell@foundry.com>