[Core] Partial HostSession C++ migration
Closes #417. Add (partial) `HostSession` implementation to C++.

This is required by the more interesting manager API methods. We omit
logging at this point as it can easily be added later, and it keeps the
slice thinner - so we can get to the harder problems such as the vector
+ variant types for resolve/register et. al.

Inherit from bound C++ `HostSession` in Python to provide the
non-migrated logging methods. Full migration will be done as part of
future work.

Use a `shared_ptr` holder type for the `Host`, in both C++ and Python,
since conceptually there is only one host. Even if technically `Host` is
simple enough to be treated with value semantics in C++, doing so raises
issues with Python (e.g. Python subclass is lost when going into
C++ then back out to Python, and tests are trickier).

Signed-off-by: David Feltell <david.feltell@foundry.com>