[Core] Move manager state freeze/thaw to `Session`
In preparation for the removal of the Transaction API (see #421), we
need to ensure there is still a way to persist or distribute a context
between processes. This was previously handled by the
`TransactionCoordinator` as a 'one stop shop for grouped actions'.

Even though we're going to remove that API for now (until we have more
concrete use cases to validate its design again), we still need to be
able to share a context across multiple processes.

As the `Session` class is already responsible for creating contexts,
moving the freeze/thaw logic there seemed sensible in the interim.

We may also want to rename these methods as there is no real reason
not to continue using a context that has been frozen for distribution
in the originating process. We can tackle this later though.

Signed-off-by: Tom Cowland <tom@foundry.com>