[Core] Remove Session.host()
Part of #331.

`Session.host()` is unused and if anything should return a
`HostInterface` rather than `Host`, since `Session` is a host-side
structure.

These changes create a problem of testability. We wish to ensure the
correct host is provided to managers via the `HostSession` passed along
to (most) `ManagerInterface` methods.

This was achieved somewhat in the tests for the test harness by using
`Session.host()`, where we just had to trust that this `Host` instance
was the same as the one in `HostSession.host()`.

Instead, for the test harness tests, store the host's identifier in
the `StubManager.info()` dict when `StubManager.initialize()` is called.
We can then assert on that in the unit tests.

Also, add an assertion to the `Session` unit tests to extract the
`HostSession` provided to the `Manager` and assert that it contains a
`Host` wrapping the expected `HostInterface`.

Signed-off-by: David Feltell <david.feltell@foundry.com>