[Core] Remove Host.py
Part of #331. Closes #427.

Remove remaining pure Python `Host` implementation. Nothing in the main
codebase uses the methods that were left. Some of this functionality
will return once we redesign auditing to work with C++.

Since the `Host` has been migrated to C++, and since this class is
non-polymoprphic, inheriting from it will not work as expected. So make
this explicit by marking as `final` in both C++ and pybind.

Signed-off-by: David Feltell <david.feltell@foundry.com>