[Core] Change `Context.kOther` to `Context.kUnknown`
Saying 'other' was misleading, as it suggested there could be some other
kind of access. Its intended use was more for the situation where the
type of access was not known.

Changing the default context access to unknown is a more accurate
starting point and will help debug hosts that aren't properly setting
the access in their code (we could easily add an audit check in
`Manager` that any unknown access contexts supplied to an API call are
flagged).

Part of #291.

Signed-off-by: Tom Cowland <tom@foundry.com>