[Core] Short-circuit persistence token generation with unsupported manager
Ensures the Manager does not attempt to invoke the interfaces
persistence token methods when no state was provided for the context.

This ensures that host code does not need to concern itself with the
specifics of whether or not a manager implements custom state management
which is an optional feature.

Signed-off-by: Tom Cowland <tom@foundry.com>