[Core] Don't create a child state if the context has no state
Ensures the Manager does not attempt to create a child state object if
the manager never provided a state in the first place.

This ensures that host code does not need to concern itself with the
specifics of whether or not a manager implements custom state management
which is an optional feature.

Signed-off-by: Tom Cowland <tom@foundry.com>