[Tests] Perftest batch resolve strategy
Part of #456. To inform our decision on whether batch-first functions
should return an array of result objects or instead execute provided
callbacks, we need to have some confidence in the relative performance
of the two approaches.

The test cases themselves cover
* Returning a `vector` of `variant<Data, Error>`-like result objects.
* Calling success/error `std::function` callbacks on each result.
  - As a separate test, also have those callbacks append to a `vector`
    to simulate the direct-to-`vector` return value case.
* Calling success/error `std::function` callbacks with a non-trivial
  capture, to bust the small-size optimization of (GCC's)
  `std::function` implementation.
* Calling success/error raw function pointer callbacks.

So update the `perftest` target to test these new scenarios. The
previous perftest is overwritten, since we expect git archaeology can
surface the previous implementation if desired for any reason. The new
implementation is still rough around the edges, but suffices for our
purposes.

The `main` takes several command-line arguments to control the number of
test iterations (epochs), warm-up epochs (see below), database size to
draw from, number of references to query in a batch, and the probability
that a reference will result in an error (not found in database).

The perftest is split into an executable and a shared library, to add a
hard barrier across which the compiler cannot do stripping or inlining
that may skew tests.

The harness is expanded to allow executing then dumping the results of
multiple test cases and the relative ratio of their timings, with the
end result being the mean and standard deviation of these relative
ratios. This functionality is gathered into a `Benchmark` struct.

A `Fixture` struct collects all the inputs required to run test cases. A
random selection from the "database" of possible inputs is chosen for
each run of each test, with a parameter determining the probability of
an erroneous input.  A new `Fixture` is created for every run of every
test.

A "warm up" of several iterations is performed before timings are taken,
in order to skip over the initial slowness caused by mysterious forces,
such as OS scheduler, CPU throttling, etc.

Signed-off-by: David Feltell <david.feltell@foundry.com>