[Core] Remove `openassetio::SharedPtr` alias
We plan to `enable_shared_from_this`, which makes less sense to alias
as it is specific to std::shared_ptr. It was raised as odd to alias one
and not the other.

We plan to properly isolate callers from the nature of the pointers in
#481 using static factory methods vs bare constructors, to ensure we can
still refactor their implementation in future if needed.

Closes #460.

Signed-off-by: Tom Cowland <tom@foundry.com>