[Core] Require `::make` for reference-semantic classes
In order to ensure consistent use of shared pointers as handles for our
reference-semantic objects, we protect their constructors and provide a
static factory method that should be used for all instantiations.

This is motivated by the complex multi-language environment in which
the API operates. Enforcing the use of a common handle (which just so
happens to be a `std::shared_ptr` at preset), means that we have a
predictable and manageable lifecycle model that works for all
combinations of hosts and managers across the language matrix.

We settled on `make` for the factory method as `create` is already in
use (`Manager.createContext`, `Specification.create`) and `new` is
reserved.

TraitsData also has a copy-constructor. We elected to migrate this to
An equivalent `make` method, that specifically takes a shared ptr over
value. The motivation for this is that there is no reason in normal use
that anyone would have access to a TraitsData instance though anything
other than a shared pointer. Using this as the input ensures that the
encapsulation of handle plus factory method is maintained, and the
handle is the sole currency through which TraitsData is traded.

This approach maps to python well, where conceptually TraitsData behaves
like a dictionary.

This does however raise a topic of how best to handle the overloaded
equality operator so that similar consistency is maintained. #493 has
been raised to track this.

Closes #481

Signed-off-by: Tom Cowland <tom@foundry.com>