[Test] Increase openassetio-python CTest verbosity
Add the `-v` flag to the pytest invocation executed when building the
`openassetio-python-pytest` CMake target (or equivalently, when running
the `openassetio-python` CTest target wrapper).

This causes `pytest` to print the individual test functions as it
executes them in a list, making it much easier to visually dissect,
especially when interleaved with output from the manager test harness
`unittest` tests.

Signed-off-by: David Feltell <david.feltell@foundry.com>