[Core] Common header for trait collection aliases
In preparation for using `TraitSet` and other trait-related collections
in #458 and beyond, create a dedicated `collection` header to put
declarations of such types.

Previously, `TraitSet` only existed as a child of `TraitsData`, but will
be used independently, and widely.

Similarly, other lists, such as of `TraitSet` and `TraitsData`, will be
used extensively throughout the batch-first API.

By using a dedicated header we can avoid unnecessary `#include`s of
complete definitions, relying instead on forward declarations.

Signed-off-by: David Feltell <david.feltell@foundry.com>