[Core] Redesign ManagerInterface initialization
Closes #503. Refactor the `setSettings`/`initialize` behaviour to remove
some of the current ambiguities.

Settings and initialization are deliberately de-coupled to facilitate
the following workflow:

1. A host queries all registered managers to obtain their display name
   (managers are required to be cheap to construct for this query).
2. A user determines which manager to use
3. A host instantiates the manager of choice with its identifier.
4. It queries the manager's default settings with getSettings so that
   the user can customize them before using the manager (eg: to
   determine which server should be used).
5. It updates the manager's settings based on a users changes.
6. It initializes the manager, where non-trivial setup operations
   happen.

This split of instantiation of a manager for general queries and
initialization is critical as initialization may be non-trivial, and so
initializing with default settings just to ask for a display name may
perform significant redundant work, especially if many manager plugins
are present. It may also be impossible to initialize a manager without
certain user-supplied settings and so this needs to be a conscious step.

The current mechanism is functional, but certain usage requirements are
less than obvious, specifically:

* It is not clear that a host must call initialize after calling
  setSettings, even if it has previously been called.
* The manager is required to buffer the settings, so that changes as a
  result of new values in setSettings don't take effect until after
  initialize has been (re)called.

The revised design makes this clear by forcing a host to call
initialize(settings) if it wishes to update them, and avoids the need
for a manager to juggle not-yet-initialized settings updates.

The existing workflow is still supported:

1. A host queries all registered managers to obtain their display name.
2. A user determines which manager to use.
3. A host instantiates the manager of choice with its identifier.
4. It queries the manager's default settings with settings.
5. It initializes the manager with the required settings, where
   non-trivial setup operations happen.

So update `ManagerInterface`/`Manager.initialize` to take an
`InfoDictionary` of settings.

Hence remove `setSettings` and migrate code to instead call
`initialize`. This then entails renaming `getSettings` to `settings`
to match our coding style re. accessor/mutator pairs.

Migrate `ManagerInterface`/`Manager.settings` to C++ to allow for C++
hosts/managers to query/provide the current settings, so they can be
updated before passing to `initialize`.

Since a settings `dict` is now required, ensure `Session` defaults the
settings to an empty `dict`, rather than `None`.

Update `apiComplianceSuite` to merge `setSettings` tests with
(new) `initialize` tests, and roll in some improvements:
* Ensure settings are not updated when given an empty dict.
* Ensure settings are not updated when only some keys are invalid.
* Remove redundant test of "some_settings_with_valid_keys", since
  "some_settings_with_all_keys" already covers it.

Update BAL to support an empty library path, so that `initialize` tests
in `apiComplianceSuite` don't need to load a library JSON file.

Fix up some `@needsref`s now that `@ref`s are available.

Signed-off-by: David Feltell <david.feltell@foundry.com>

[Core] Migrate `settings` to C++

Part of #503. Migrate `ManagerInterface`/`Manager.settings` to C++. This
is a necessary part of allowing a C++ host to configure a manager before
instantiation. This then prepares the way for `initialize` to accept
settings from both Python and C++.

Signed-off-by: David Feltell <david.feltell@foundry.com>

[Core] Rename `getSettings` to `settings`

In preparation for the redesign of `ManagerInterface` initialization in
#503, rename `ManagerInterface`/`Manager.getSettings` to `settings`.

Signed-off-by: David Feltell <david.feltell@foundry.com>